Fix tests instability by removing parallelization
Secure storage is a point that makes test unstable. Indeed, in a
parallele execution scheme, some of them store information in it,
others remove it. It breachs the assumptions on which they rely.
Making them sequential is not the cleanest and finest way to fix
that, but it's a simple working one.